--- Backend/HealthApi/src/main/java/com/teamOne/healthApi/entities/Profesional.java
@@ -1,4 +1,0 @@
-package com.teamOne.healthApi.entities;
-
-public class Profesional {
-}
