+++ proyectoSalud/src/main/java/com/equipoUno/proyectoSalud/entities/Profesional.java
@@ -1,0 +1,4 @@
+package com.equipoUno.proyectoSalud.entities;
+
+public class Profesional {
+}
