+++ proyectoSalud/src/main/resources/application.properties
