+++ proyectoSalud/src/test/java/com/equipoUno/proyectoSalud/HealthApiApplicationTests.java
@@ -1,4 +1,4 @@
-package com.teamOne.healthApi;
+package com.equipoUno.proyectoSalud;
 
 import org.junit.jupiter.api.Test;
 import org.springframework.boot.test.context.SpringBootTest;
